requirements.txt
Djangj==1.5.2
PIL==1.1.7
South==0.8.2	
django-grappelli==2.4.6
django-mptt==0.6.0
django-photologue==2.6.1
django-tagging==0.3.1
django-treeadmin==0.4
